+++ packages/muelu/src/Utils/MueLu_TimeMonitor.hpp
@@ -69,69 +69,16 @@ namespace MueLu {
 
   public:
 
-    TimeMonitor(const BaseClass& object, const std::string& msg, MsgType timerLevel = Timings0) {
-      // Inherit props from 'object'
-      SetVerbLevel      (object.GetVerbLevel());
-      SetProcRankVerbose(object.GetProcRankVerbose());
+    TimeMonitor(const BaseClass& object, const std::string& msg, MsgType timerLevel = Timings0);
 
-      if (IsPrint(timerLevel) &&
-          /* disable timer if never printed: */ (IsPrint(RuntimeTimings) || (!IsPrint(NoTimeReport)))) {
-
-        if (!IsPrint(NoTimeReport)) {
-          // TODO: there is no function to register a timer in Teuchos::TimeMonitor after the creation of the timer. But would be useful...
-          timer_ = Teuchos::TimeMonitor::getNewTimer("MueLu: " + msg);
-        } else {
-          timer_ = rcp(new Teuchos::Time("MueLu: " + msg));
-        }
-
-        // Start the timer (this is what is done by Teuchos::TimeMonitor)
-        timer_->start();
-        timer_->incrementNumCalls();
-#ifdef HAVE_TEUCHOS_ADD_TIME_MONITOR_TO_STACKED_TIMER
-        const auto stackedTimer = Teuchos::TimeMonitor::getStackedTimer();
-        if (nonnull(stackedTimer))
-          stackedTimer->start(timer_->name());
-#endif
-      }
-    }
-
-    ~TimeMonitor() {
-      // Stop the timer if present
-      if (timer_ != Teuchos::null) {
-        timer_->stop();
-#ifdef HAVE_TEUCHOS_ADD_TIME_MONITOR_TO_STACKED_TIMER
-        try {
-          const auto stackedTimer = Teuchos::TimeMonitor::getStackedTimer();
-          if (nonnull(stackedTimer))
-            stackedTimer->stop(timer_->name());
-        }
-        catch (std::runtime_error) {
-          std::ostringstream warning;
-          warning <<
-            "\n*********************************************************************\n"
-            "WARNING: Overlapping timers detected!\n"
-            "A TimeMonitor timer was stopped before a nested subtimer was\n"
-            "stopped. This is not allowed by the StackedTimer. This corner case\n"
-            "typically occurs if the TimeMonitor is stored in an RCP and the RCP is\n"
-            "assigned to a new timer. To disable this warning, either fix the\n"
-            "ordering of timer creation and destuction or disable the StackedTimer\n"
-            "support in the TimeMonitor by setting the StackedTimer to null\n"
-            "with:\n"
-            "Teuchos::TimeMonitor::setStackedTimer(Teuchos::null)\n"
-            "*********************************************************************\n";
-          std::cout << warning.str() << std::endl;
-          Teuchos::TimeMonitor::setStackedTimer(Teuchos::null);
-        }
-#endif
-      }
-    }
+    ~TimeMonitor();
 
   protected:
-    TimeMonitor() { }
+    TimeMonitor();
 
   private:
     RCP<Teuchos::Time> timer_;
-  };
+  }; //class TimeMonitor
 
   //TODO: code duplication MutuallyExclusiveTimeMonitor / TimeMonitor
 
@@ -182,6 +129,9 @@ namespace MueLu {
   private:
     RCP<MutuallyExclusiveTime<TagName> > timer_; // keep a reference on the timer to print stats if RuntimeTimings=ON //TODO:use base class instead
   };
+
+  extern template class MutuallyExclusiveTimeMonitor<FactoryBase>;
+  extern template class MutuallyExclusiveTimeMonitor<Level>;
 
 } // namespace MueLu
